created Customer and Employee (extends Person) Employee has 3 subclasses Vehicule + ref to Owner
//TODO
* check how jobs are referenced (id or object)
* methods to assign jobs to mechanic from manager

+++ src/main/java/com/somena/employee/Manager.java
@@ -5,18 +5,19 @@ import java.util.List;
 
 public class Manager extends Employee{
 
-    private List<Integer> jobList;
+    //assuming jobs are ref by ID which is  Long
+    private List<Long> jobList;
 
-    public Manager(String firstName, String lastName, long id, Date startDate, List<Integer> jobList) {
+    public Manager(String firstName, String lastName, long id, Date startDate, List<Long> jobList) {
         super(firstName, lastName, id, startDate);
         this.jobList = jobList;
     }
 
-    public List<Integer> getJobList() {
+    public List<Long> getJobList() {
         return jobList;
     }
 
-    public void setJobList(List<Integer> jobList) {
+    public void setJobList(List<Long> jobList) {
         this.jobList = jobList;
     }
 }
